pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jira-rag"
version = "0.1.0"
description = "RAG indexer and search API over Jira issues, comments, and linked merge requests"
requires-python = ">=3.11"
dependencies = [
    "atlassian-python-api>=3.41.0",
    "qdrant-client>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "voyageai>=0.3.0",
    "fastembed>=0.4.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "structlog>=24.0.0",
    "tenacity>=9.0.0",
    "httpx>=0.28.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
jira-rag = "jira_rag.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/jira_rag"]
